feat: add team-based player visibility control; refactor visibility tracker to support external updates
- Implement setAllTeamplayerVisible to set visibility for all players in a team based on advantage state
- Refactor PlayerVisibilityTracker to allow static visibility updates and external control
- Update AdvancementListener to use new visibility logic instead of scoreboard commands
- Remove unused imports and minor code cleanups in related classes

+++ src/client/java/com/rcutanf/teamhunter/client/ui/TeamScoreHud.java
@@ -1,7 +1,6 @@
 package com.rcutanf.teamhunter.client.ui;
 
 import com.rcutanf.teamhunter.Phase;
-import com.rcutanf.teamhunter.client.TeamhunterClient;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.font.TextRenderer;
 import net.minecraft.client.gui.DrawContext;
@@ -328,9 +327,9 @@ public class TeamScoreHud {
         int textHeight = textRenderer.fontHeight;
 
         context.getMatrices().push();
-        context.getMatrices().translate(x + textWidth / 2, y + textHeight / 2, 0);
+        context.getMatrices().translate(x + (float) textWidth / 2, y + (float) textHeight / 2, 0);
         context.getMatrices().scale(scale, scale, 1.0f);
-        context.getMatrices().translate(-(x + textWidth / 2), -(y + textHeight / 2), 0);
+        context.getMatrices().translate(-(x + (float) textWidth / 2), -(y + (float) textHeight / 2), 0);
 
         context.drawText(textRenderer, text, x, y, color, false);
